test(generate-speech): Add tests for generateSpeech()
