+++ lib/generate-speech.js
@@ -62,7 +62,7 @@ exports.callAws = (info, i, callback) => {
     .replace(/(X-Amz-Credential=)([^&]*)/, '$1...')
     .replace(/(X-Amz-Signature=.)([^&]*)/, '$1...')
   debug('callAws')(`Making request to ${sanitizedUrl}`)
-  got.stream(url).on('error', err => { error = err }).pipe(outputStream)
+  got.stream(url).on('error', /* istanbul ignore next */ err => { error = err }).pipe(outputStream)
 }
 
 /**
@@ -124,7 +124,6 @@ exports.generateSpeech = (ctx, task) => {
   const strParts = ctx.parts
 
   // Add in the default options.
-  /* istanbul ignore next */
   ctx.opts = Object.assign({}, {
     'access-key': ctx.args.accessKey,
     ffmpeg: ctx.args.ffmpeg || 'ffmpeg',
@@ -137,21 +136,16 @@ exports.generateSpeech = (ctx, task) => {
     type: ctx.args.type || 'text',
     voice: ctx.args.voice || 'Joanna'
   }, ctx.args)
-  /* istanbul ignore next */
   if (typeof ctx.opts.lexicon !== 'undefined' && !Array.isArray(ctx.opts.lexicon)) {
     ctx.opts.lexicon = [ctx.opts.lexicon]
   }
-  /* istanbul ignore next */
   debug('generateSpeech')(`Options: ${JSON.stringify(sanitizeOpts(ctx.opts))}`)
 
-  /* istanbul ignore next */
   let polly = exports.createPolly(ctx.opts)
 
   // Compile the text parts and options together in a packet.
-  /* istanbul ignore next */
   let parts = strParts.map(part => exports.buildInfo(part, polly.getSynthesizeSpeechUrl.bind(polly), task, ctx.opts))
 
-  /* istanbul ignore next */
   return exports.generateAll(parts, ctx.opts, exports.callAws, task)
     .then(exports.createManifest)
     .then(manifest => {
